requirements.txt
mysql-connector-python==8.3.0
matplotlib==3.7.5
pandas==2.1.3


This file allows you to install all dependencies at once using:

pip install -r requirements.txt



